+++ packages/shuvi/src/shuvi/shuvi.dev.ts
@@ -5,7 +5,6 @@ import { acceptsHtml } from '../lib/utils';
 import { serveStatic } from '../lib/serveStatic';
 import Base, { IShuviConstructorOptions } from './shuvi.base';
 import { throwServerRenderError } from '../lib/throw';
-import { matchPath } from '@shuvi/router';
 
 export default class ShuviDev extends Base {
   private _onDemandRouteMgr!: OnDemandRouteManager;
@@ -95,26 +94,7 @@ export default class ShuviDev extends Base {
     return async (ctx, next) => {
       const middlewares = this._getServerMiddlewares();
 
-      let i = 0;
-
-      const runMiddleware = async (
-        middleware: Runtime.IServerMiddlewareModule
-      ) => {
-        if (i === middlewares.length) {
-          await next();
-          return;
-        }
-        const matchedPath = matchPath(middleware.path, ctx.request.url);
-        if (!matchedPath) {
-          await runMiddleware(middlewares[++i]);
-          return;
-        }
-        ctx.params = matchedPath.params;
-
-        await middleware.handler(ctx, () => runMiddleware(middlewares[++i]));
-      };
-
-      await runMiddleware(middlewares[i]);
+      await this._runServerMiddlewares(middlewares)(ctx, next);
     };
   };
 }
